backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
moviepy==1.0.3
requests
google-auth-oauthlib
google-api-python-client
supabase
Pillow
numpy
ayrshare